Fix demo.ipynb
Fixed the "Demo the LLM" section of the demo.ipynb notebook to
work correctly when your config requires the llama server to be
started manually.

Simplified the LLAMA3 and CHATML templates.

Fixed a typo in glados.py

Increased sleep time between attempts from .01 to .50 in llama.py
because sometimes it takes more than 0.10 seconds for llama-server
to even start.