More straightforward exit handling
Use `on_unmount` method to shutdown gracefully, instead of overriding `action_quit`.  `on_unmount` is called whenever the app exits (even on an exception) and handles more shutdown cases. It is the best place to put code which must run on exit.  Also, overriding `action_quit` leads to type checking errors, because the method signature is not the same as in the super class.